% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nclRadvanced-package.R
\docType{package}
\name{jrAdvanced-package}
\alias{jrAdvanced-package}
\alias{jrAdvanced}
\title{R courses by Jumping Rivers}
\description{
Functions and datasets used in for R 
courses by Jumping Rivers. This package is used in
the "Advanced R programming" course".
}
\keyword{package}
