% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_args.R
\name{print_dots}
\alias{print_dots}
\title{Print the ... argument}
\usage{
print_dots(...)
}
\arguments{
\item{...}{The ... argument from the function to be analysed}
}
\description{
A really simple function that explores what happens when use ...
}
